agents/requirements.txt
deepagents==0.1.4
langchain-anthropic==0.3.15
langchain-google-genai==2.1.3
langchain==0.3.25
langchain-core==0.3.57
langgraph==0.4.1
fastapi==0.115.6
uvicorn[standard]==0.34.0
pillow==11.1.0
pydantic==2.10.6
httpx==0.28.1
python-dotenv==1.0.1
